moved inject headers logic to common utils
