+++ packages/core/http/core-http-router-server-internal/src/versioned_router/core_versioned_route.ts
@@ -38,7 +38,7 @@ import {
   readVersion,
   removeQueryVersion,
 } from './route_version_utils';
-import { injectResponseHeaders } from './inject_response_headers';
+import { injectResponseHeaders } from '../util';
 import { validRouteSecurity } from '../security_route_config_validator';
 
 import { resolvers } from './handler_resolvers';
